+++ app/services/data/export/fetch-table.service.js
@@ -5,7 +5,7 @@
  * @module FetchTableService
  */
 
-const { db } = require('../../../db/db.js')
+const { db } = require('../../../../db/db.js')
 
 /**
  * Retrieves headers, a knex query for the table rows and the table name from the table in the db, and returns them as an object
