+++ test/services/data/export/delete-files.service.test.js
@@ -14,7 +14,7 @@ const path = require('path')
 const mockFs = require('mock-fs')
 
 // Thing under test
-const DeleteFilesService = require('../../../app/services/db-export/delete-files.service.js')
+const DeleteFilesService = require('../../../../app/services/data/export/delete-files.service')
 
 describe('Delete Files service', () => {
   let filenameWithPath
